Milestone_3/requirements.txt
mlflow==2.9.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
PyYAML==6.0.1
apache-airflow==2.8.1
